requirements.txt
streamlit==1.32.0
requests==2.31.0
python-dotenv==1.0.0
pillow==10.1.0
pydantic==1.10.8 